re-merging cli file to reduce PR size
